feat n refactor: nambahin BE buat prediksi diagnose, buat code dummy buat IoT, mindahin component, mindahin data ke lib/data
