+++ apps/web/src/components/pages/login/InputField.tsx
@@ -1,0 +1,59 @@
+"use client"
+
+import type React from "react"
+
+import { useState } from "react"
+import { Eye, EyeOff } from "lucide-react"
+import { cn } from "@/lib/utils"
+
+interface InputFieldProps extends React.InputHTMLAttributes<HTMLInputElement> {
+  label: string
+  error?: string
+}
+
+export function InputField({ id, label, type = "text", error, className, ...props }: InputFieldProps) {
+  const [showPassword, setShowPassword] = useState(false)
+  const isPassword = type === "password"
+
+  const togglePasswordVisibility = () => {
+    setShowPassword(!showPassword)
+  }
+
+  return (
+    <div className="space-y-1">
+      <label htmlFor={id} className="block text-sm font-medium text-gray-700">
+        {label}
+      </label>
+      <div className="relative">
+        <input
+          id={id}
+          type={isPassword && showPassword ? "text" : type}
+          className={cn(
+            "w-full px-4 py-2 bg-white text-gray-800 border rounded-md focus:outline-none focus:ring-2",
+            error ? "border-red-500 focus:ring-red-500" : "border-gray-300 focus:ring-green-500",
+            className,
+          )}
+          aria-invalid={error ? "true" : "false"}
+          aria-describedby={error ? `${id}-error` : undefined}
+          {...props}
+        />
+        {isPassword && (
+          <button
+            type="button"
+            onClick={togglePasswordVisibility}
+            className="absolute right-3 top-1/2 transform -translate-y-1/2 text-gray-500 hover:text-gray-700 focus:outline-none"
+            aria-label={showPassword ? "Hide password" : "Show password"}
+            tabIndex={-1}
+          >
+            {showPassword ? <EyeOff size={18} /> : <Eye size={18} />}
+          </button>
+        )}
+      </div>
+      {error && (
+        <p id={`${id}-error`} className="text-sm text-red-600">
+          {error}
+        </p>
+      )}
+    </div>
+  )
+}
