+++ apps/web/src/lib/data/data-komoditas.ts
@@ -1,0 +1,8 @@
+export const KOMODITAS = [
+    "Bawang Merah",
+    "Bawang Putih Bonggol",
+    "Beras Medium",
+    "Beras Premium",
+    "Cabai Merah Keriting",
+    "Cabai Rawit Merah",
+  ];
