+++ apps/web/src/lib/data/komoditas-deskripsi.ts
@@ -1,0 +1,14 @@
+export const DESKRIPSI: Record<string, string> = {
+    "bawang-merah":
+      "Bawang Merah adalah salah satu bumbu dapur utama di Indonesia, digunakan untuk berbagai masakan dan memberikan cita rasa khas.",
+    "bawang-putih-bonggol":
+      "Bawang Putih Bonggol adalah bawang putih utuh yang sering digunakan sebagai bumbu dasar berbagai masakan Indonesia.",
+    "beras-medium":
+      "Beras Medium adalah beras dengan kualitas menengah, cocok untuk konsumsi sehari-hari dengan harga yang lebih terjangkau.",
+    "beras-premium":
+      "Beras Premium adalah beras dengan kualitas terbaik, kadar air rendah, dan butir utuh. Cocok untuk konsumsi sehari-hari.",
+    "cabai-merah-keriting":
+      "Cabai Merah Keriting adalah cabai dengan bentuk panjang dan keriting, memiliki rasa pedas dan sering digunakan dalam sambal.",
+    "cabai-rawit-merah":
+      "Cabai Rawit Merah adalah cabai kecil berwarna merah dengan tingkat kepedasan tinggi, sering digunakan untuk menambah rasa pedas pada masakan.",
+  };
